move 'db' to top-level directory
