+++ src/app/controllers/OptionTransactionController.ts
@@ -1,4 +1,4 @@
-import * as DB from '../../db';
+import * as DB from '../../../db';
 import * as Types from '../../lib/types';
 
 export class OptionTransactionController {
